feat: Phase 8 — Embedded maintenance operations (rebuild, compact, WAL prune) (#80)
* docs(08): capture phase context

* docs(state): record phase 8 context session

* docs(08): research embedded maintenance phase domain

* docs(08-embedded-maintenance): create phase plan

Two plans in 2 waves:
- Plan 01: Expand EmbeddedSession with callback slots, wire JNA + Panama backends
- Plan 02: Integration tests for embedded maintenance in both backends

* fix(08): revise plans based on checker feedback

Add Java lint step (gradle :core:check :jna:check :panama:check) to
08-01 verification and Task 2 verify block per CLAUDE.md compliance.

Document D-09 data-seeded test deferral in 08-02 objective and task
actions — collection CRUD unavailable until FUTURE-03.

* docs(08): add validation strategy

* feat(08-01): expand EmbeddedSession with callback slots and maintenance methods

- Add 5 BiFunction callback slots for maintenance operations
- Add 7 public maintenance methods (5 primary + 2 convenience overloads)
- Add ensureOpen() guard pattern matching ServerSession
- Constructor expands from 2 to 7 parameters with null checks

* feat(08-01): wire JNA and Panama backends with maintenance symbol bindings

- Add 5 chroma_embedded_* symbol declarations to JNA JnaBindings interface
- Add 5 MethodHandle fields to Panama Ffi record with eager binding
- Both doStartEmbedded methods pass callFfiJson lambdas to expanded constructor
- Fix EmbeddedSessionTest to use 7-parameter constructor

* docs(08-01): complete embedded maintenance wiring plan

- Create 08-01-SUMMARY.md with execution results
- Update STATE.md with phase 8 position and decisions
- Update ROADMAP.md progress (1/2 plans complete)
- Mark EMNT-01 through EMNT-04 complete in REQUIREMENTS.md

* test(08-02): add JNA embedded maintenance integration tests

- 7 tests covering smoke, error paths, input validation, closed-session guards
- Verifies rebuildCollection, compactAll, compactCollection, pruneAllWAL, pruneCollectionWAL

* test(08-02): add Panama embedded maintenance integration tests

- 7 tests mirroring JNA coverage for smoke, error paths, validation, closed-session guards
- Verifies rebuildCollection, compactAll, compactCollection, pruneAllWAL, pruneCollectionWAL

* docs(08-02): complete embedded maintenance integration tests plan

- SUMMARY.md with 14 tests (7 JNA + 7 Panama) coverage
- STATE.md updated with phase 08 completion
- ROADMAP.md marks phase 08 complete (2/2 plans)
- REQUIREMENTS.md marks EMNT-05 complete

* docs(phase-08): complete phase execution

* docs(phase-08): evolve PROJECT.md after phase completion

* test(08): complete UAT - 5 passed, 0 issues

* docs(08): ship phase 08 — PR #80

* fix(08): address PR review findings — dead name param, error handling, tests

- Remove dead `name` parameter from rebuildCollection and pruneCollectionWAL
  (options object already carries the name; old API silently discarded it)
- Add pruneCollectionWAL(String) convenience overload for API parity
- Log readLastError failures in PanamaChromaRuntime instead of silent null
- Use addSuppressed for callFfiVoid drain catch in AbstractChromaRuntime
- Replace null-returning test stubs with UnsupportedOperationException
- Add 13 unit tests: constructor null-rejection (5), input validation (5),
  convenience overload delegation (2), existing lifecycle (6 unchanged)
- Update integration tests and ServerSession stubs for new signatures

* fix(08): address PR review — stale stubs, type consistency, API symmetry

- Update 3 stale "Phase 8" stub messages to "Phase 10" in ServerSession
- Add compactCollection(String) convenience overload to match rebuild/prune
- Add Javadoc on pruneAllWAL noting name field is ignored
- Change collectionCount from u32/int to u64/long across Rust, Go, Java
- Change WAL prune warning from Option<String>/String to Vec<String>/List

* test(08): add compactCollection delegation test and convenience overload smoke tests

Complete the symmetric delegation test set in EmbeddedSessionTest and
exercise the convenience overload FFI path in JNA/Panama integration tests.

+++ internal/runtime/compaction.go
@@ -39,7 +39,7 @@ type CompactionCollectionResult struct {
 // CompactionResult captures explicit compaction execution metadata.
 type CompactionResult struct {
 	// CollectionCount is the number of collections attempted, including entries with per-collection errors.
-	CollectionCount       uint32                       `json:"collection_count"`
+	CollectionCount       uint64                       `json:"collection_count"`
 	DurationMS            uint64                       `json:"duration_ms"`
 	PendingOpsBeforeTotal uint64                       `json:"pending_ops_before_total"`
 	PendingOpsAfterTotal  uint64                       `json:"pending_ops_after_total"`
